// platform-frontend/src/infraestructure/interfaces/guides.response.ts
export interface GuideResponse {
    id:             string;
    title:          string;
    description:    string;
    filePath:       string;
    uploadDate:     string;
    gradeId:        string; 
    gradeName:      string;
    subjectId:      string;
    subjectName:    string;
    uploadedByName: string;
}